+++ test/unit/verify/verify.test.js
@@ -1,5 +1,5 @@
 const crypto = require('crypto')
-jest.mock('../../app/config/verify', () => ({ totalRetries: 1 }))
+jest.mock('../../../app/config/verify', () => ({ totalRetries: 1 }))
 
 const ORIGINAL_CTL_BATCH_FILE_NAME = 'CTL_PENDING_TEST_BATCH.dat'
 
@@ -88,7 +88,7 @@ jest.mock('@azure/storage-blob', () => {
     }
   }
 })
-const verify = require('../../app/verify')
+const verify = require('../../../app/verify')
 
 describe('verification', () => {
   const setupFileContent = (isValid) => {
